[WIP][package] Add package dependency visualizer
**Summary**
Visualizing the dependency graph of a package can be useful for
debugging. This commit adds a dependency graph visualizer based on
Cytoscape, a Javascript library for rendering graphs. The visualization
is created as an HTML file that can be opened in a browser. This HTML
page supports zooming in/out, dragging nodes around,
collapsing/expanding modules, and collapsing expanding dependencies of a
module.

**Test Plan**
See the attached video.

ghstack-source-id: 775e88a0565b6b3a6f08a85d4339e693ceae4d6d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54242